requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.35.0
JPype1==1.4.1